projects/sie-pipeline/requirements.txt
httpx
python-dotenv
supabase
